# Optimize `sort!` for AbstractArray